Revamp regcurly(), regpiece() use of it
This commit copies portions of new_regcurly(), which has been around
since 5.28, into plain regcurly(), as a baby step in preparation for
converting entirely to the new one.  These functions are used for
parsing {m,n} quantifiers.  The new version will add capabilities not
in the old.

The commit adds an optional parameter to return to the caller,
information it gleans during parsing about where the bounds are for the
numbers in the quantifier string.

This commit changes regpiece() to use this information, instead of
itself reparsing (which could have led to things being out-of-sync), and
places the bounds handling in a separate function, instead of having
near duplicate code for the lower and upper bounds